requirements.txt
numpy==1.24.2
pandas==2.0.0
matplotlib==3.7.1
python-dotenv
pyautogen
azure-identity
seaborn
scikit-learn
requests
urllib3
nltk
pillow
pytest
beautifulsoup4
jupyter